+++ src/physics.rs
@@ -1,6 +1,6 @@
 use super::math::*;
 pub trait Physics {
-    fn get_m(&self) -> f64;
+    fn get_m(&self) -> f32;
     fn get_v(&self) -> &Vector2;
     fn set_v(&mut self, v: &Vector2) -> ();
     fn get_p(&self) -> &Vector2;
